feat(router): also provide NavigationOptions in Activation events
Especially at the start of the application, the NavigationStarted event may be missed and
the application cannot get the NavigationOptions. So, we just continue propagting the
options along the lifecycle of the navigation until the end.

This will also make the application data in the options available at each lifecycle stage.