feat: reference the owning docker
The docker is needed to act on the docks and dock groups. Rather than having
to pass it around, it is more intuitive to just get it from one of the objects it
manages.

The docker reference is updated anytime the object becomes managed by the
docker (i.e. docked in any form).