fix: update CommunityToolkit.Mvvm and fix windows SDK error
see https://github.com/CommunityToolkit/dotnet/issues/923

As soon as the new update for the .NET SDK is released next month, this will just work with no changes needed. Manually setting the package version is a temporary workaround, but you should remove it when the new .NET SDK is out (as you don't want to hardcode a package version that will be eventually outdated).