feat: app config management module
The PathFinder moved to the DroidNet.Config module and is now a reusable
service, that can be extended to add custom paths to resolve.

A new SettingService is introduced that handles the abstraction of the
Options Pattern usage, fine grained change notifications on top of the
Ioptionsmonitor and serialization of settings to Json.