fix(resources): properly get resources using MRT Core
fix(resources): properly get resources using MRT Core

First, give a chance to the application to override the localized string
resource. If not found, try to get the localized string from the calling
assembly's sub-tree in the main resource map (Note: This will only work
if the application is a packaged app.)

Last, try to get the localized string directly from the calling
assembly's resource map.