refactor: cleanup access to the graphics backend within its module
Several entities in the graphics module were accessing the backend
directly. This commit refactors the code to ensure that all access to
the backend is done through the `GetGraphics()` internal function. This
has the advantage of avoiding the need to pass around the backend
instance almost everywhere in the code and having to store it as a
member in so many objects.