refactor: generic resource/asset loading API
BREAKING CHANGE:
- All resource-related APIs in AssetLoader now use ResourceKey instead
  of PakFile& and ResourceIndexT.

Unified loading API using ResourceKey forresources and AssetKey for
assets. The AssetLoader is the central point for loading and the only
place where content is cached.

Added initial support for refrence counting in AssetLoader, allowing
for shared ownership of resources. This is a step towards a more robust
resource management system.