fix: avoid conflicts with windows min/max macros
Although defining NOMINMAX is a common practice to avoid conflicts with
the Windows headers, it is not reliable for all scenarios (such as
static analysis for example).

The following changes ensure that the code does not conflict with the
Windows min/max macros by using std::min and std::max explicitly:

  - put parentheses around the function names of std::min and std::max
  - do the same for std::numeric_limits<T>::max() and min()

This syntax won't apply a function-like macro. (Formally, to apply a
function-like macro the name of the macro must be followed by optional
white space then a '('.)