feat(input): partial integration of InputSystem
InputSystem is now an engine module, that primarily gets involved during
the PhaseId::kInput phase. Event handling is currently still following
the legacy callback model, which should be replaced with per-frame input
snapshot.

InputSystem is integrated with the async event pump, using its
`BroadcastChannel`. This works fine for now, but in the future we might
want to consider tuning the number of events that get processed per
frame, to avoid input lag when there are many events (e.g. mouse move).
We might also want to add throttling and coalescing for certain event
types.