feat(platform): add event filtering mechanism and improve logs
The `Platform` now supports an event filtering mechanism that allows
users to set a custom callable to process events before they are
dispatched. A new async task has been introduced to handle event
filtering, and runs first in the event processing pipeline.

Only one filter can be set at a time, and setting a new filter replaces
the previous one. The filter can be removed by calling
`ClearEventFilter()`.

Filters can only decide if an event is handled or not, and cannot
modify the event itself. If an event is marked as handled by the filter,
it will not be processed further.